Added employee, record, payroll, survey and profile routes and controllers file
